Pattern Causality Lightweight Function

This function implements the Pattern Causality Model Mk. II for lightweight analysis of causal interactions between two time series using pattern and signature spaces. It assesses causality through reconstructed state spaces and hashed pattern analysis.

Usage

pcLightweight(X, Y, E, tau, metric, h, weighted, tpb = TRUE)

Arguments

X

A numeric vector representing the first time series.

Y

A numeric vector representing the second time series.

E

The embedding dimension, which influences the number of dimensions in which the time series is reconstructed for analysis.

tau

The time delay used in reconstructing the time series in the embedded space.

metric

A character string indicating the distance metric to be used (e.g., 'euclidean', 'maximum').

h

The prediction horizon, representing the number of steps ahead for which predictions are needed.

weighted

A logical indicating whether to use a weighted approach in the causality strength calculations.

tpb

A bool parameter to show progress bar.

Value

A data frame with columns for total, positive, negative, and dark causality percentages across evaluated time points, giving insights into the nature of causality between the time series.

Examples

data(climate_indices)
X <- climate_indices$AO
Y <- climate_indices$AAO
result <- pcLightweight(X, Y, E = 3, tau = 1, metric = "euclidean", h = 2, weighted = TRUE)
#>   |                                                                        |                                                                |   0%  |                                                                        |#                                                               |   2%  |                                                                        |##                                                              |   2%  |                                                                        |##                                                              |   3%  |                                                                        |##                                                              |   4%  |                                                                        |###                                                             |   4%  |                                                                        |###                                                             |   5%  |                                                                        |####                                                            |   6%  |                                                                        |####                                                            |   7%  |                                                                        |#####                                                           |   7%  |                                                                        |#####                                                           |   8%  |                                                                        |######                                                          |   9%  |                                                                        |######                                                          |  10%  |                                                                        |#######                                                         |  10%  |                                                                        |#######                                                         |  11%  |                                                                        |#######                                                         |  12%  |                                                                        |########                                                        |  12%  |                                                                        |########                                                        |  13%  |                                                                        |#########                                                       |  13%  |                                                                        |#########                                                       |  14%  |                                                                        |#########                                                       |  15%  |                                                                        |##########                                                      |  15%  |                                                                        |##########                                                      |  16%  |                                                                        |###########                                                     |  16%  |                                                                        |###########                                                     |  17%  |                                                                        |###########                                                     |  18%  |                                                                        |############                                                    |  18%  |                                                                        |############                                                    |  19%  |                                                                        |############                                                    |  20%  |                                                                        |#############                                                   |  20%  |                                                                        |#############                                                   |  21%  |                                                                        |##############                                                  |  21%  |                                                                        |##############                                                  |  22%  |                                                                        |##############                                                  |  23%  |                                                                        |###############                                                 |  23%  |                                                                        |###############                                                 |  24%  |                                                                        |################                                                |  24%  |                                                                        |################                                                |  25%  |                                                                        |################                                                |  26%  |                                                                        |#################                                               |  26%  |                                                                        |#################                                               |  27%  |                                                                        |##################                                              |  28%  |                                                                        |##################                                              |  29%  |                                                                        |###################                                             |  29%  |                                                                        |###################                                             |  30%  |                                                                        |####################                                            |  31%  |                                                                        |####################                                            |  32%  |                                                                        |#####################                                           |  32%  |                                                                        |#####################                                           |  33%  |                                                                        |######################                                          |  34%  |                                                                        |######################                                          |  35%  |                                                                        |#######################                                         |  35%  |                                                                        |#######################                                         |  36%  |                                                                        |#######################                                         |  37%  |                                                                        |########################                                        |  37%  |                                                                        |########################                                        |  38%  |                                                                        |#########################                                       |  38%  |                                                                        |#########################                                       |  39%  |                                                                        |#########################                                       |  40%  |                                                                        |##########################                                      |  40%  |                                                                        |##########################                                      |  41%  |                                                                        |###########################                                     |  41%  |                                                                        |###########################                                     |  42%  |                                                                        |###########################                                     |  43%  |                                                                        |############################                                    |  43%  |                                                                        |############################                                    |  44%  |                                                                        |#############################                                   |  45%  |                                                                        |#############################                                   |  46%  |                                                                        |##############################                                  |  46%  |                                                                        |##############################                                  |  47%  |                                                                        |##############################                                  |  48%  |                                                                        |###############################                                 |  48%  |                                                                        |###############################                                 |  49%  |                                                                        |################################                                |  49%  |                                                                        |################################                                |  50%  |                                                                        |################################                                |  51%  |                                                                        |#################################                               |  51%  |                                                                        |#################################                               |  52%  |                                                                        |##################################                              |  52%  |                                                                        |##################################                              |  53%  |                                                                        |##################################                              |  54%  |                                                                        |###################################                             |  54%  |                                                                        |###################################                             |  55%  |                                                                        |####################################                            |  56%  |                                                                        |####################################                            |  57%  |                                                                        |#####################################                           |  57%  |                                                                        |#####################################                           |  58%  |                                                                        |#####################################                           |  59%  |                                                                        |######################################                          |  59%  |                                                                        |######################################                          |  60%  |                                                                        |#######################################                         |  60%  |                                                                        |#######################################                         |  61%  |                                                                        |#######################################                         |  62%  |                                                                        |########################################                        |  62%  |                                                                        |########################################                        |  63%  |                                                                        |#########################################                       |  63%  |                                                                        |#########################################                       |  64%  |                                                                        |#########################################                       |  65%  |                                                                        |##########################################                      |  65%  |                                                                        |##########################################                      |  66%  |                                                                        |###########################################                     |  67%  |                                                                        |###########################################                     |  68%  |                                                                        |############################################                    |  68%  |                                                                        |############################################                    |  69%  |                                                                        |#############################################                   |  70%  |                                                                        |#############################################                   |  71%  |                                                                        |##############################################                  |  71%  |                                                                        |##############################################                  |  72%  |                                                                        |###############################################                 |  73%  |                                                                        |###############################################                 |  74%  |                                                                        |################################################                |  74%  |                                                                        |################################################                |  75%  |                                                                        |################################################                |  76%  |                                                                        |#################################################               |  76%  |                                                                        |#################################################               |  77%  |                                                                        |##################################################              |  77%  |                                                                        |##################################################              |  78%  |                                                                        |##################################################              |  79%  |                                                                        |###################################################             |  79%  |                                                                        |###################################################             |  80%  |                                                                        |####################################################            |  80%  |                                                                        |####################################################            |  81%  |                                                                        |####################################################            |  82%  |                                                                        |#####################################################           |  82%  |                                                                        |#####################################################           |  83%  |                                                                        |#####################################################           |  84%  |                                                                        |######################################################          |  84%  |                                                                        |######################################################          |  85%  |                                                                        |#######################################################         |  85%  |                                                                        |#######################################################         |  86%  |                                                                        |#######################################################         |  87%  |                                                                        |########################################################        |  87%  |                                                                        |########################################################        |  88%  |                                                                        |#########################################################       |  88%  |                                                                        |#########################################################       |  89%  |                                                                        |#########################################################       |  90%  |                                                                        |##########################################################      |  90%  |                                                                        |##########################################################      |  91%  |                                                                        |###########################################################     |  92%  |                                                                        |###########################################################     |  93%  |                                                                        |############################################################    |  93%  |                                                                        |############################################################    |  94%  |                                                                        |#############################################################   |  95%  |                                                                        |#############################################################   |  96%  |                                                                        |##############################################################  |  96%  |                                                                        |##############################################################  |  97%  |                                                                        |##############################################################  |  98%  |                                                                        |############################################################### |  98%  |                                                                        |############################################################### |  99%  |                                                                        |################################################################|  99%  |                                                                        |################################################################| 100%
print(result)
#>       total  positive  negative      dark
#> 1 0.3140187 0.2371795 0.3846154 0.3782051