Projected Nearest Neighbors Information

Extracts and returns information about the projected nearest neighbors in a time series context, specifically useful for understanding interactions in dynamic complex systems.

Usage

projectedNNsInfo(My, Dy, SMy, PSMy, timesX, i, h)

Arguments

My

Matrix of coordinates in the original space.

Dy

Distance matrix, representing distances between elements in My.

SMy

Matrix of signatures, capturing essential patterns in the data.

PSMy

Matrix of patterns, representing characteristic configurations of the data.

timesX

Index at which the projection starts.

i

Index of the specific element for which information is being extracted.

h

Horizon over which the projection is considered.

Value

A list containing indices of the element, the projected times, distances, weights derived from these distances, signatures, patterns, and the coordinates of the nearest neighbors.

Examples

set.seed(123)
E <- 3
tau <- 1
Mx <- matrix(rnorm(200), nrow = 20)
My <- matrix(rnorm(200), nrow = 20)
Dx <- distanceMatrix(Mx, "minkowski")
Dy <- distanceMatrix(My, "minkowski")
SMx <- signatureSpace(Mx, E)
SMy <- signatureSpace(My, E)
PSMx <- patternSpace(SMx, E)
PSMy <- patternSpace(SMy, E)
CCSPAN <- (E - 1) * tau
NNSPAN <- E + 1
i <- 15
h <- 2
NNx <- pastNNsInfo(CCSPAN, NNSPAN, Mx, Dx, SMx, PSMx, i, h)
timesX <- NNx$times
projNNy <- projectedNNsInfo(My, Dy, SMy, PSMy, timesX, i, h)
print(projNNy)
#> $i
#> [1] 15
#> 
#> $Times_Projected
#> [1]  4 10 12  3
#> 
#> $Dists
#>        4       10       12        3 
#> 4.336187 3.795081 4.405915 4.136853 
#> 
#> $Weights
#> [1] 0.2474734 0.2556362 0.2464407 0.2504497
#> 
#> $Signatures
#>            [,1]       [,2]       [,3]        [,4]       [,5]       [,6]
#> [1,]  0.1643943 -1.8448920  2.4375023 -1.56585030 -0.7868617  1.7133336
#> [2,] -0.9615645  1.5010526 -0.3962528  0.07885816  1.8695086 -2.0302839
#> [3,] -0.3340070 -0.4715568  1.5244043  0.09165305 -0.2974063 -1.9452088
#> [4,]  1.3749932  0.3862125 -1.4489062  0.33887214 -1.3245653 -0.4002356
#>            [,7]       [,8]        [,9]
#> [1,] -0.6048036  0.9686565 -2.07969028
#> [2,] -0.1871963 -0.2811321  0.80714196
#> [3,] -0.8043145  0.3416996 -0.50659744
#> [4,]  3.0098291 -1.3972883 -0.09904009
#> 
#> $Patterns
#> [1]  51293202 123799878  51213990  44681970
#> 
#> $Coordinates
#>             [,1]        [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,]  0.54319406  0.70758835 -1.1373036 1.30019868 -0.2656516 -1.0525133
#> [2,] -0.05402813 -1.01559258  0.4854600 0.08920722  0.1680654  2.0375740
#> [3,]  0.24368743 -0.09031959 -0.5618764 0.96252797  1.0541810  0.7567748
#> [4,] -0.26514506  1.10984814  1.4960607 0.04715443  0.3860266 -0.9385387
#>             [,7]        [,8]       [,9]      [,10]
#> [1,]  0.66082030  0.05601673  1.0246732 -1.0550170
#> [2,]  0.00729009 -0.17990623 -0.4610383  0.3461036
#> [3,] -1.18843404 -1.99274849 -1.6510489 -2.1576463
#> [4,] -1.33877429  1.67105483  0.2737665  0.1747264
#>