Create Signature Space from Embedded Matrix

This function computes the signature space of a matrix obtained from a time series' state space by calculating the successive differences of each embedded vector. The signature space reveals changes between successive points in the time series, capturing dynamics that are crucial for understanding complex system behaviors.

Usage

signatureSpace(M, E)

Arguments

M

Matrix, the embedded state space matrix where each row is a point in the reconstructed state space of the time series.

E

Integer, the embedding dimension used to create the matrix M.

Value

Matrix where each row represents the vector of differences between successive elements of the corresponding row in matrix M. The orientation of the matrix may vary depending on the embedding dimension.

Examples

data(climate)
ts <- climate$AO
E <- 3
tau <- 1
stateSpaceMatrix <- stateSpace(ts, E, tau)
signatureMatrix <- signatureSpace(stateSpaceMatrix, E)
print(signatureMatrix)
#>           [,1]    [,2]
#>   [1,]  1.5361 -0.1174
#>   [2,] -0.1174 -0.3427
#>   [3,] -0.3427  0.9067
#>   [4,]  0.9067  1.1833
#>   [5,]  1.1833 -0.8947
#>   [6,] -0.8947 -0.7226
#>   [7,] -0.7226  0.6382
#>   [8,]  0.6382 -1.1975
#>   [9,] -1.1975  1.7185
#>  [10,]  1.7185  0.8197
#>  [11,]  0.8197 -3.3605
#>  [12,] -3.3605  1.1320
#>  [13,]  1.1320 -0.4996
#>  [14,] -0.4996  1.0142
#>  [15,]  1.0142 -0.7357
#>  [16,] -0.7357  1.8763
#>  [17,]  1.8763 -1.3437
#>  [18,] -1.3437  0.4370
#>  [19,]  0.4370  0.4978
#>  [20,]  0.4978 -0.8338
#>  [21,] -0.8338 -0.8398
#>  [22,] -0.8398  1.3037
#>  [23,]  1.3037 -0.0590
#>  [24,] -0.0590 -0.2153
#>  [25,] -0.2153 -1.3131
#>  [26,] -1.3131  2.0751
#>  [27,]  2.0751 -0.2508
#>  [28,] -0.2508 -0.6175
#>  [29,] -0.6175  0.9984
#>  [30,]  0.9984 -0.8046
#>  [31,] -0.8046 -0.7960
#>  [32,] -0.7960 -0.1274
#>  [33,] -0.1274  0.9798
#>  [34,]  0.9798 -1.0280
#>  [35,] -1.0280  0.3323
#>  [36,]  0.3323  1.8573
#>  [37,]  1.8573  0.1002
#>  [38,]  0.1002  0.3797
#>  [39,]  0.3797 -1.6625
#>  [40,] -1.6625 -0.9714
#>  [41,] -0.9714  1.1849
#>  [42,]  1.1849  0.3574
#>  [43,]  0.3574  0.1955
#>  [44,]  0.1955 -0.7687
#>  [45,] -0.7687  0.8719
#>  [46,]  0.8719  0.3063
#>  [47,]  0.3063  0.3919
#>  [48,]  0.3919 -3.1650
#>  [49,] -3.1650  1.2388
#>  [50,]  1.2388 -0.1707
#>  [51,] -0.1707  0.2969
#>  [52,]  0.2969  0.7534
#>  [53,]  0.7534 -0.1820
#>  [54,] -0.1820  0.9673
#>  [55,]  0.9673 -0.9309
#>  [56,] -0.9309  1.2020
#>  [57,]  1.2020 -2.0568
#>  [58,] -2.0568  0.8741
#>  [59,]  0.8741  0.7188
#>  [60,]  0.7188 -1.2077
#>  [61,] -1.2077 -2.0833
#>  [62,] -2.0833  2.1024
#>  [63,]  2.1024  0.7628
#>  [64,]  0.7628 -0.4719
#>  [65,] -0.4719  0.0116
#>  [66,]  0.0116  0.4468
#>  [67,]  0.4468 -0.8785
#>  [68,] -0.8785  0.1425
#>  [69,]  0.1425 -0.6956
#>  [70,] -0.6956  1.4119
#>  [71,]  1.4119 -3.2517
#>  [72,] -3.2517  1.3659
#>  [73,]  1.3659  1.9912
#>  [74,]  1.9912  0.1010
#>  [75,]  0.1010 -1.0846
#>  [76,] -1.0846  0.0856
#>  [77,]  0.0856 -0.0430
#>  [78,] -0.0430  0.3882
#>  [79,]  0.3882  0.1158
#>  [80,]  0.1158  0.9207
#>  [81,]  0.9207 -2.2526
#>  [82,] -2.2526 -0.7301
#>  [83,] -0.7301  1.3800
#>  [84,]  1.3800 -2.3365
#>  [85,] -2.3365  4.8349
#>  [86,]  4.8349 -1.8278
#>  [87,] -1.8278  0.2639
#>  [88,]  0.2639  0.1677
#>  [89,]  0.1677 -0.5429
#>  [90,] -0.5429 -0.8180
#>  [91,] -0.8180  0.8029
#>  [92,]  0.8029  1.4480
#>  [93,]  1.4480 -0.4989
#>  [94,] -0.4989 -0.8659
#>  [95,] -0.8659 -1.2074
#>  [96,] -1.2074 -0.3256
#>  [97,] -0.3256 -0.2733
#>  [98,] -0.2733  2.1335
#>  [99,]  2.1335 -0.0618
#> [100,] -0.0618 -1.0355
#> [101,] -1.0355  0.2440
#> [102,]  0.2440 -0.3694
#> [103,] -0.3694  1.1222
#> [104,]  1.1222 -0.3665
#> [105,] -0.3665 -0.4558
#> [106,] -0.4558  0.0019
#> [107,]  0.0019  0.7986
#> [108,]  0.7986 -1.3309
#> [109,] -1.3309  0.8691
#> [110,]  0.8691 -0.3636
#> [111,] -0.3636 -0.2854
#> [112,] -0.2854  0.9066
#> [113,]  0.9066 -0.2039
#> [114,] -0.2039  0.3980
#> [115,]  0.3980  0.7847
#> [116,]  0.7847 -1.0069
#> [117,] -1.0069 -0.0671
#> [118,] -0.0671  1.7135
#> [119,]  1.7135  1.4272
#> [120,]  1.4272  0.1733
#> [121,]  0.1733 -1.7490
#> [122,] -1.7490 -1.7805
#> [123,] -1.7805  1.1390
#> [124,]  1.1390 -0.5438
#> [125,] -0.5438  0.5206
#> [126,]  0.5206 -0.3147
#> [127,] -0.3147  0.1522
#> [128,]  0.1522  0.2876
#> [129,]  0.2876 -0.9569
#> [130,] -0.9569 -0.6775
#> [131,] -0.6775  1.6444
#> [132,]  1.6444  2.4009
#> [133,]  2.4009 -0.4116
#> [134,] -0.4116 -1.1112
#> [135,] -1.1112 -0.9360
#> [136,] -0.9360 -0.6385
#> [137,] -0.6385 -0.6001
#> [138,] -0.6001  0.1156
#> [139,]  0.1156 -0.0302
#> [140,] -0.0302  0.8707
#> [141,]  0.8707 -0.1397
#> [142,] -0.1397  0.7561
#> [143,]  0.7561 -0.5535
#> [144,] -0.5535 -1.5992
#> [145,] -1.5992  0.3492
#> [146,]  0.3492  1.0570
#> [147,]  1.0570 -0.0437
#> [148,] -0.0437 -0.6019
#> [149,] -0.6019 -0.0726
#> [150,] -0.0726  0.9849
#> [151,]  0.9849 -0.9091
#> [152,] -0.9091 -0.1397
#> [153,] -0.1397  0.5366
#> [154,]  0.5366  1.3285
#> [155,]  1.3285 -1.0632
#> [156,] -1.0632  0.5717
#> [157,]  0.5717 -0.1375
#> [158,] -0.1375 -1.5047
#> [159,] -1.5047  1.8619
#> [160,]  1.8619 -1.6434
#> [161,] -1.6434  0.4931
#> [162,]  0.4931  0.3442
#> [163,]  0.3442 -1.1756
#> [164,] -1.1756  0.2744
#> [165,]  0.2744  1.0829
#> [166,]  1.0829  0.9097
#> [167,]  0.9097  1.8686
#> [168,]  1.8686 -3.3108
#> [169,] -3.3108  0.5798
#> [170,]  0.5798 -1.1997
#> [171,] -1.1997 -1.1721
#> [172,] -1.1721  1.0880
#> [173,]  1.0880  0.0088
#> [174,]  0.0088  0.1177
#> [175,]  0.1177  0.0324
#> [176,]  0.0324 -0.2044
#> [177,] -0.2044  1.5668
#> [178,]  1.5668 -1.1059
#> [179,] -1.1059 -0.1838
#> [180,] -0.1838 -0.5736
#> [181,] -0.5736  2.7425
#> [182,]  2.7425 -1.6563
#> [183,] -1.6563 -0.3401
#> [184,] -0.3401  1.7217
#> [185,]  1.7217 -1.2556
#> [186,] -1.2556  0.4768
#> [187,]  0.4768 -0.9116
#> [188,] -0.9116  0.2581
#> [189,]  0.2581  1.6054
#> [190,]  1.6054 -0.8856
#> [191,] -0.8856 -1.0476
#> [192,] -1.0476  1.5827
#> [193,]  1.5827 -1.0357
#> [194,] -1.0357 -1.3563
#> [195,] -1.3563  0.0719
#> [196,]  0.0719  0.7794
#> [197,]  0.7794 -0.1053
#> [198,] -0.1053  0.7607
#> [199,]  0.7607 -1.0926
#> [200,] -1.0926  0.6240
#> [201,]  0.6240 -0.7983
#> [202,] -0.7983 -1.4038
#> [203,] -1.4038  0.9267
#> [204,]  0.9267  1.3636
#> [205,]  1.3636 -1.6464
#> [206,] -1.6464 -0.0419
#> [207,] -0.0419  1.2987
#> [208,]  1.2987  0.7231
#> [209,]  0.7231  0.2179
#> [210,]  0.2179 -0.5899
#> [211,] -0.5899 -1.2647
#> [212,] -1.2647  1.3225
#> [213,]  1.3225 -0.0461
#> [214,] -0.0461 -1.8572
#> [215,] -1.8572  1.2640
#> [216,]  1.2640  2.3455
#> [217,]  2.3455 -0.7979
#> [218,] -0.7979 -0.7672
#> [219,] -0.7672 -1.2847
#> [220,] -1.2847  0.1461
#> [221,]  0.1461  0.3844
#> [222,]  0.3844  0.5512
#> [223,]  0.5512  0.0739
#> [224,]  0.0739 -0.8942
#> [225,] -0.8942  0.0386
#> [226,]  0.0386  0.5900
#> [227,]  0.5900 -2.0095
#> [228,] -2.0095  1.8974
#> [229,]  1.8974 -0.0712
#> [230,] -0.0712  0.2165
#> [231,]  0.2165  0.4665
#> [232,]  0.4665 -1.1393
#> [233,] -1.1393  0.4990
#> [234,]  0.4990  0.8620
#> [235,]  0.8620 -1.7002
#> [236,] -1.7002  1.3442
#> [237,]  1.3442 -1.7437
#> [238,] -1.7437  2.8028
#> [239,]  2.8028 -1.2431
#> [240,] -1.2431  0.3718
#> [241,]  0.3718 -1.9737
#> [242,] -1.9737  1.7760
#> [243,]  1.7760 -0.0585
#> [244,] -0.0585  0.4811
#> [245,]  0.4811 -0.7090
#> [246,] -0.7090 -0.6701
#> [247,] -0.6701  0.7312
#> [248,]  0.7312 -0.0649
#> [249,] -0.0649  0.6167
#> [250,]  0.6167  0.4322
#> [251,]  0.4322  0.2271
#> [252,]  0.2271 -0.1944
#> [253,] -0.1944 -1.5272
#> [254,] -1.5272  0.1729
#> [255,]  0.1729  1.2476
#> [256,]  1.2476 -0.3830
#> [257,] -0.3830 -1.2355
#> [258,] -1.2355  0.7933
#> [259,]  0.7933  0.2510
#> [260,]  0.2510 -0.0781
#> [261,] -0.0781 -1.8983
#> [262,] -1.8983 -0.7729
#> [263,] -0.7729  1.3956
#> [264,]  1.3956  0.3364
#> [265,]  0.3364 -1.0641
#> [266,] -1.0641  2.5925
#> [267,]  2.5925 -0.4540
#> [268,] -0.4540 -0.4673
#> [269,] -0.4673 -0.0157
#> [270,] -0.0157  0.5515
#> [271,]  0.5515 -1.2271
#> [272,] -1.2271  1.4141
#> [273,]  1.4141  0.1111
#> [274,]  0.1111 -2.1410
#> [275,] -2.1410  2.7037
#> [276,]  2.7037 -0.0778
#> [277,] -0.0778 -0.4015
#> [278,] -0.4015 -0.1536
#> [279,] -0.1536 -0.3470
#> [280,] -0.3470  0.1713
#> [281,]  0.1713 -0.2451
#> [282,] -0.2451 -0.5561
#> [283,] -0.5561  0.1858
#> [284,]  0.1858 -1.4458
#> [285,] -1.4458  0.0634
#> [286,]  0.0634 -0.1670
#> [287,] -0.1670  1.1204
#> [288,]  1.1204  0.5995
#> [289,]  0.5995  0.8052
#> [290,]  0.8052 -1.1111
#> [291,] -1.1111  1.1948
#> [292,]  1.1948 -1.1188
#> [293,] -1.1188  0.1774
#> [294,]  0.1774 -0.3557
#> [295,] -0.3557  0.7470
#> [296,]  0.7470 -1.1364
#> [297,] -1.1364  1.3122
#> [298,]  1.3122 -0.3772
#> [299,] -0.3772 -1.9510
#> [300,] -1.9510  0.1573
#> [301,]  0.1573  1.8466
#> [302,]  1.8466 -0.7275
#> [303,] -0.7275  0.3151
#> [304,]  0.3151 -0.1416
#> [305,] -0.1416  0.0354
#> [306,]  0.0354 -0.5197
#> [307,] -0.5197  1.5752
#> [308,]  1.5752 -1.3704
#> [309,] -1.3704  1.1937
#> [310,]  1.1937  0.5518
#> [311,]  0.5518 -0.8739
#> [312,] -0.8739 -1.6268
#> [313,] -1.6268 -0.0773
#> [314,] -0.0773  1.3017
#> [315,]  1.3017 -0.7172
#> [316,] -0.7172  0.3802
#> [317,]  0.3802  0.3530
#> [318,]  0.3530  0.0563
#> [319,]  0.0563  0.7763
#> [320,]  0.7763 -0.7726
#> [321,] -0.7726  0.1979
#> [322,]  0.1979 -2.3316
#> [323,] -2.3316  1.9334
#> [324,]  1.9334  0.0147
#> [325,]  0.0147 -1.4480
#> [326,] -1.4480  1.7421
#> [327,]  1.7421  0.0175
#> [328,]  0.0175  0.9150
#> [329,]  0.9150 -0.9681
#> [330,] -0.9681 -0.3679
#> [331,] -0.3679  0.8717
#> [332,]  0.8717 -1.6356
#> [333,] -1.6356  1.5504
#> [334,]  1.5504  1.7604
#> [335,]  1.7604 -0.2479
#> [336,] -0.2479 -3.3407
#> [337,] -3.3407  2.4890
#> [338,]  2.4890 -0.6378
#> [339,] -0.6378  0.3494
#> [340,]  0.3494 -1.4487
#> [341,] -1.4487  0.1585
#> [342,]  0.1585  0.3628
#> [343,]  0.3628  0.2126
#> [344,]  0.2126  0.2046
#> [345,]  0.2046 -0.9022
#> [346,] -0.9022  1.3398
#> [347,]  1.3398 -0.0021
#> [348,] -0.0021  0.1191
#> [349,]  0.1191 -0.3525
#> [350,] -0.3525 -1.0406
#> [351,] -1.0406 -0.7497
#> [352,] -0.7497  1.1149
#> [353,]  1.1149 -0.3901
#> [354,] -0.3901  0.3994
#> [355,]  0.3994 -0.2461
#> [356,] -0.2461  2.0024
#> [357,]  2.0024 -1.5836
#> [358,] -1.5836  0.5556
#> [359,]  0.5556  0.1519
#> [360,]  0.1519 -1.4720
#> [361,] -1.4720  0.7936
#> [362,]  0.7936  0.8512
#> [363,]  0.8512  0.2212
#> [364,]  0.2212 -2.5444
#> [365,] -2.5444 -0.0052
#> [366,] -0.0052  1.3022
#> [367,]  1.3022  0.9282
#> [368,]  0.9282 -2.4144
#> [369,] -2.4144  1.9988
#> [370,]  1.9988 -3.8717
#> [371,] -3.8717  0.8260
#> [372,]  0.8260 -1.6789
#> [373,] -1.6789  3.8336
#> [374,]  3.8336  0.1576
#> [375,]  0.1576 -0.6441
#> [376,] -0.6441  0.9056
#> [377,]  0.9056  0.4483
#> [378,]  0.4483 -0.5519
#> [379,] -0.5519 -0.7480
#> [380,] -0.7480  0.3976
#> [381,]  0.3976  0.0913
#> [382,]  0.0913 -2.2553
#> [383,] -2.2553  0.9479
#> [384,]  0.9479  3.2585
#> [385,]  3.2585 -0.1513
#> [386,] -0.1513  0.8507
#> [387,]  0.8507 -2.3099
#> [388,] -2.3099 -0.8227
#> [389,] -0.8227  0.3862
#> [390,]  0.3862 -0.5910
#> [391,] -0.5910  1.7273
#> [392,]  1.7273  0.1351
#> [393,]  0.1351  0.6594
#> [394,]  0.6594  0.7616
#> [395,]  0.7616 -2.4405
#> [396,] -2.4405  0.1834
#> [397,]  0.1834  1.0734
#> [398,]  1.0734 -1.0717
#> [399,] -1.0717  0.2030
#> [400,]  0.2030 -0.8408
#> [401,] -0.8408  0.8402
#> [402,]  0.8402 -0.1538
#> [403,] -0.1538  0.7582
#> [404,]  0.7582 -2.2862
#> [405,] -2.2862  1.4034
#> [406,]  1.4034 -1.6380
#> [407,] -1.6380  1.1391
#> [408,]  1.1391 -0.3979
#> [409,] -0.3979 -2.1780
#> [410,] -2.1780  3.5076
#> [411,]  3.5076  0.1718
#> [412,]  0.1718  0.0547
#> [413,]  0.0547 -0.5598
#> [414,] -0.5598  0.1653
#> [415,]  0.1653 -0.6151
#> [416,] -0.6151  0.7237
#> [417,]  0.7237  1.7662
#> [418,]  1.7662 -0.5541
#> [419,] -0.5541 -2.4437
#> [420,] -2.4437  1.0126
#> [421,]  1.0126  1.1620
#> [422,]  1.1620 -0.2338
#> [423,] -0.2338 -0.5078
#> [424,] -0.5078 -0.9716
#> [425,] -0.9716  0.0185
#> [426,]  0.0185  0.1174
#> [427,]  0.1174  0.4734
#> [428,]  0.4734 -1.2363
#> [429,] -1.2363  0.6041
#> [430,]  0.6041  0.9432
#> [431,]  0.9432  0.6787
#> [432,]  0.6787 -0.0490
#> [433,] -0.0490  0.7948
#> [434,]  0.7948 -0.6217
#> [435,] -0.6217 -0.4529
#> [436,] -0.4529 -0.3358
#> [437,] -0.3358 -1.5349
#> [438,] -1.5349  0.4189
#> [439,]  0.4189  0.5245
#> [440,]  0.5245 -0.0856
#> [441,] -0.0856  2.1951
#> [442,]  2.1951 -0.5009
#> [443,] -0.5009 -2.8928
#> [444,] -2.8928  1.4252
#> [445,]  1.4252  0.3037
#> [446,]  0.3037 -1.3313
#> [447,] -1.3313  1.0154
#> [448,]  1.0154  0.3486
#> [449,]  0.3486 -0.2281
#> [450,] -0.2281  0.3876
#> [451,]  0.3876  0.3086
#> [452,]  0.3086 -2.6983
#> [453,] -2.6983  1.3064
#> [454,]  1.3064  2.3973
#> [455,]  2.3973 -0.8444
#> [456,] -0.8444 -0.6021
#> [457,] -0.6021  1.0255
#> [458,]  1.0255 -1.4541
#> [459,] -1.4541 -0.6414
#> [460,] -0.6414  1.1318
#> [461,]  1.1318  0.2325
#> [462,]  0.2325 -0.4842
#> [463,] -0.4842 -0.6424
#> [464,] -0.6424  1.1827
#> [465,]  1.1827 -0.7679
#> [466,] -0.7679  0.0186
#> [467,]  0.0186 -0.2218
#> [468,] -0.2218  0.3935
#> [469,]  0.3935 -1.0538
#> [470,] -1.0538  1.4850
#> [471,]  1.4850  0.6357
#> [472,]  0.6357 -0.7997
#> [473,] -0.7997  0.2319
#> [474,]  0.2319  0.2243
#> [475,]  0.2243 -0.2516
#> [476,] -0.2516 -0.1717
#> [477,] -0.1717 -1.5290
#> [478,] -1.5290  1.2259
#> [479,]  1.2259 -0.8229
#> [480,] -0.8229  1.8627
#> [481,]  1.8627  0.9666
#> [482,]  0.9666 -2.3714
#> [483,] -2.3714 -0.9760
#> [484,] -0.9760  0.6300
#> [485,]  0.6300 -0.2884
#> [486,] -0.2884  0.1679
#> [487,]  0.1679  1.0280
#> [488,]  1.0280 -0.3884
#> [489,] -0.3884 -1.1112
#> [490,] -1.1112  1.6055
#> [491,]  1.6055  2.0069
#> [492,]  2.0069  0.9982
#> [493,]  0.9982 -0.7758
#> [494,] -0.7758 -1.7133
#> [495,] -1.7133 -0.9552
#> [496,] -0.9552 -0.0947
#> [497,] -0.0947 -0.2900
#> [498,] -0.2900  0.0306
#> [499,]  0.0306  1.0126
#> [500,]  1.0126 -0.7031
#> [501,] -0.7031  2.1581
#> [502,]  2.1581 -3.8224
#> [503,] -3.8224 -0.7476
#> [504,] -0.7476  1.2929
#> [505,]  1.2929  3.2999
#> [506,]  3.2999 -2.3136
#> [507,] -2.3136  0.0438
#> [508,]  0.0438  1.0052
#> [509,]  1.0052 -0.2144
#> [510,] -0.2144 -0.8395
#> [511,] -0.8395 -0.0423
#> [512,] -0.0423  0.1058
#> [513,]  0.1058  0.2388
#> [514,]  0.2388  0.1051
#> [515,]  0.1051  0.6502
#> [516,]  0.6502  0.6961
#> [517,]  0.6961 -1.2392
#> [518,] -1.2392 -0.9078
#> [519,] -0.9078  1.8261
#> [520,]  1.8261 -1.2977
#> [521,] -1.2977  0.0991
#> [522,]  0.0991 -0.1949
#> [523,] -0.1949 -0.4852
#> [524,] -0.4852  2.0009
#> [525,]  2.0009 -1.0068
#> [526,] -1.0068 -3.0581
#> [527,] -3.0581  2.0449
#> [528,]  2.0449  2.2747
#> [529,]  2.2747 -1.3201
#> [530,] -1.3201 -1.2534
#> [531,] -1.2534  2.1074
#> [532,]  2.1074 -1.4205
#> [533,] -1.4205  0.1317