Summarize Pattern Causality Effect

Provides a summary of the pattern causality effect analysis results. This function displays the summary statistics for the effects, including the number of components and the strongest effects.

Usage

# S3 method for class 'pc_effect'
summary(object, ...)

Arguments

object

A pc_effect object.

...

Additional arguments passed to the summary function.

Value

Invisibly returns the input object.

Examples

# \donttest{
data(climate_indices)
dataset <- climate_indices[, -1]
pc_matrix_obj <- pcMatrix(dataset, E = 3, tau = 1, 
  metric = "euclidean", h = 1, weighted = TRUE, 
  verbose = FALSE)
effects <- pcEffect(pc_matrix_obj)
summary(effects)
#> $effects_summary
#> $effects_summary$positive
#>       received   exerted          Diff
#> mean 83.202119  83.20212 -3.552714e-15
#> sd    8.764969  19.95980  1.993609e+01
#> min  71.506182  64.57246 -2.649411e+01
#> max  92.589225 109.64732  2.098342e+01
#> 
#> $effects_summary$negative
#>       received  exerted      Diff
#> mean 77.082125 77.08212   0.00000
#> sd    4.887857 19.58312  22.25668
#> min  71.124192 50.33220 -18.46447
#> max  82.707750 97.20834  32.37555
#> 
#> $effects_summary$dark
#>        received    exerted      Diff
#> mean 139.715757 139.715757   0.00000
#> sd     6.942656   7.968076  11.83639
#> min  134.139041 130.631078 -13.70569
#> max  149.749949 149.992272  11.53075
#> 
#> 
#> $n_components
#> [1] 4
#> 
#> $strongest_effects
#> $strongest_effects[[1]]
#> $strongest_effects[[1]]$component
#> [1] "AAO"
#> 
#> $strongest_effects[[1]]$effect
#> [1] -26.49411
#> 
#> 
#> $strongest_effects[[2]]
#> $strongest_effects[[2]]$component
#> [1] "AAO"
#> 
#> $strongest_effects[[2]]$effect
#> [1] 32.37555
#> 
#> 
#> $strongest_effects[[3]]
#> $strongest_effects[[3]]$component
#> [1] "AO"
#> 
#> $strongest_effects[[3]]$effect
#> [1] -13.70569
#> 
#> 
#> 
#> attr(,"class")
#> [1] "summary.pc_effect"
# }